refactor: extract keyboard zoom into createZoom hook
Separate user-preference concern (font size persistence + key capture)
from terminal rendering concern. Terminal now consumes fontSize as a
reactive signal via createEffect instead of owning the state, localStorage,
and keydown listener directly.
